rename files to snake case to avoid namespace collisions
