+++ aamva_barcode_file/subfile_designator.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import NamedTuple
 
-from .FileHeader import FileHeader
+from .file_header import FileHeader
 
 
 class SubfileDesignator(NamedTuple):
