+++ tests/test_Subfile.py
@@ -1,11 +1,11 @@
 import pytest
 
-import aamva_barcode_file.Subfile as Subfile
+import aamva_barcode_file.subfile as subfile
 
 
 testdata_ids = ("1 element", "multi element", "multi element with empty")
 testdata = (
-    # ((test_string, designator, subfile), ...)
+    # ((test_string, designator, expects), ...)
     (
         "DLAAA123\r",
         ("DL", 0, 9),
@@ -24,16 +24,16 @@ testdata = (
 )
 
 
-@pytest.mark.parametrize("test_string, designator, subfile", testdata, ids=testdata_ids)
-def test_can_parse_subfile(test_string, designator, subfile):
-    assert Subfile.Subfile.parse(test_string, Subfile.SubfileDesignator(*designator)) == Subfile.Subfile(*subfile)
+@pytest.mark.parametrize("test_string, designator, expects", testdata, ids=testdata_ids)
+def test_can_parse_subfile(test_string, designator, expects):
+    assert subfile.Subfile.parse(test_string, subfile.SubfileDesignator(*designator)) == subfile.Subfile(*expects)
 
 
 def test_parse_raises_value_error_on_missing_subfile_type():
     with pytest.raises(ValueError, match="Subfile is missing subfile type."):
-        Subfile.Subfile.parse("DDAAA123\r", Subfile.SubfileDesignator("DL", 0, 9))
+        subfile.Subfile.parse("DDAAA123\r", subfile.SubfileDesignator("DL", 0, 9))
 
 
 def test_parse_raises_value_error_on_missing_segment_terminator():
     with pytest.raises(ValueError, match="Subfile is missing segment terminator."):
-        Subfile.Subfile.parse("DLAAA123", Subfile.SubfileDesignator("DL", 0, 8))
+        subfile.Subfile.parse("DLAAA123", subfile.SubfileDesignator("DL", 0, 8))
